+++ backend/src/main/java/toonpick/app/controller/AdminReviewController.java
@@ -1,23 +1,25 @@
-package toonpick.app.controller.review;
+package toonpick.app.controller;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import toonpick.app.dto.WebtoonReviewDTO;
 import toonpick.app.service.WebtoonReviewService;
-import toonpick.app.dto.PagedResponseDTO;
 
 @RequiredArgsConstructor
 @RestController
-@RequestMapping("/api/public/reviews")
-public class PublicWebtoonReviewController {
+@RequestMapping("/api/admin/reviews")
+public class AdminReviewController {
 
     private final WebtoonReviewService webtoonReviewService;
+
+    // todo : Admin 사용 가능한 API
+    // todo : 추후 @PreAuthorize("hasRole('ADMIN')") 어노테이션을 추가할 것
 
     // 특정 리뷰 조회
     @GetMapping("/{reviewId}")
@@ -26,14 +28,10 @@ public class PublicWebtoonReviewController {
         return reviewDTO != null ? ResponseEntity.ok(reviewDTO) : ResponseEntity.status(HttpStatus.NOT_FOUND).build();
     }
 
-    // 특정 웹툰의 리뷰들 가져오기
-    @GetMapping("/webtoon/{webtoonId}")
-    public ResponseEntity<PagedResponseDTO<WebtoonReviewDTO>> getReviewsByWebtoon(
-            @PathVariable Long webtoonId,
-            @RequestParam(defaultValue = "latest") String sortBy,
-            @RequestParam(defaultValue = "0") int page,
-            @RequestParam(defaultValue = "20") int size) {
-        PagedResponseDTO<WebtoonReviewDTO> pagedReviews = webtoonReviewService.getReviewsByWebtoon(webtoonId, sortBy, page, size);
-        return ResponseEntity.ok(pagedReviews);
+    // 리뷰 강제 삭제 (관리자 권한)
+    @DeleteMapping("/{reviewId}")
+    public ResponseEntity<Void> deleteReview(@PathVariable Long reviewId) {
+        webtoonReviewService.deleteReview(reviewId);
+        return ResponseEntity.noContent().build();
     }
 }
