+++ backend/src/main/java/toonpick/app/controller/MemberController.java
@@ -1,4 +1,4 @@
-package toonpick.app.controller.member;
+package toonpick.app.controller;
 
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
