+++ backend/src/main/java/toonpick/app/controller/WebtoonDataUpdateController.java
@@ -1,4 +1,4 @@
-package toonpick.app.controller.webtoon;
+package toonpick.app.controller;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
